Initial changes for DLL support.
The CMake build system has been modified to produce library-specific
headers with various macros supporting DLL creation.  The major
difference between a DLL on Windows and a .so on Linux (at least when
developing) is ensuring that appropriate code is "exported" for client
use.  This is done implicitly in .so creation, and the code actually
needs to be modified for doing this exporting.  This is implemented
using architecture-dependent macros facilitated quite a bit by built-in
CMake macros.

+++ src/angle/ChebyshevDPN.cc
@@ -26,7 +26,7 @@ ChebyshevDPN::ChebyshevDPN(const size_t dim,
 
   // Chebyshev points are just equally spaced phi's in the first quadrant
   using detran_utilities::pi;
-  for (int i = 0; i < na; ++i)
+  for (size_t i = 0; i < na; ++i)
   {
     d_phi[i] = 0.25 * (2*i + 1)*pi / na;
     d_cos_phi[i] = std::cos(d_phi[i]);
@@ -46,7 +46,7 @@ ChebyshevDPN::ChebyshevDPN(const size_t dim,
   generate_gl_parameters(np, x, w);
 
   // fill array
-  for (int i = 0; i < np; ++i)
+  for (size_t i = 0; i < np; ++i)
   {
     // Put in order.
     size_t j = np - i - 1;
