+++ src/angle/ChebyshevLegendre.cc
@@ -25,7 +25,7 @@ ChebyshevLegendre::ChebyshevLegendre(const size_t dim,
 
   // Chebyshev points are just equally spaced phi's in the first quadrant
   using detran_utilities::pi;
-  for (int i = 0; i < na; ++i)
+  for (size_t i = 0; i < na; ++i)
   {
     d_phi[i] = 0.25 * (2*i + 1)*pi / na;
     d_cos_phi[i] = std::cos(d_phi[i]);
@@ -45,7 +45,7 @@ ChebyshevLegendre::ChebyshevLegendre(const size_t dim,
   generate_gl_parameters(2*np, x, w);
 
   // fill array
-  for (int i = 0; i < np; ++i)
+  for (size_t i = 0; i < np; ++i)
   {
     size_t j = np - i - 1;
     d_cos_theta[j]    = x[i];
