+++ src/kinetics/SyntheticMomentSource.i.hh
@@ -65,12 +65,12 @@ inline void SyntheticMomentSource::build(const double dt,
   double a_0 = bdf_coefs[order-1][0];
 
   // Clear the source
-  for (int g = 0; g < d_material->number_groups(); ++g)
-    for (int cell = 0; cell < d_mesh->number_cells(); ++cell)
+  for (size_t g = 0; g < d_material->number_groups(); ++g)
+    for (size_t cell = 0; cell < d_mesh->number_cells(); ++cell)
       d_source[g][cell] = 0.0;
 
   // Add all backward terms
-  for (int j = 0; j < order; ++j)
+  for (size_t j = 0; j < order; ++j)
   {
     Assert(states[j]);
     if (size_precursor)
@@ -81,12 +81,12 @@ inline void SyntheticMomentSource::build(const double dt,
     // Skip the first entry, which is for the (n+1) term
     double a_j = bdf_coefs[order-1][j + 1];
 
-    for (int g = 0; g < d_material->number_groups(); ++g)
+    for (size_t g = 0; g < d_material->number_groups(); ++g)
     {
 
       // Add the flux term
       double phi_factor = a_j / dt / d_material->velocity(g);
-      for (int cell = 0; cell < d_mesh->number_cells(); ++cell)
+      for (size_t cell = 0; cell < d_mesh->number_cells(); ++cell)
       {
 //        std::cout << " phi factor=" << phi_factor
 //                  << " phi = " << states[j]->phi(g)[cell]
@@ -97,11 +97,11 @@ inline void SyntheticMomentSource::build(const double dt,
       // Add the precursor concentration, if applicable
       if (size_precursor)
       {
-        for (int i = 0; i < np; ++i)
+        for (size_t i = 0; i < np; ++i)
         {
           double C_factor = a_j * d_material->lambda(i) /
                             (a_0 + dt * d_material->lambda(i));
-          for (int cell = 0; cell < d_mesh->number_cells(); ++cell)
+          for (size_t cell = 0; cell < d_mesh->number_cells(); ++cell)
           {
 
             d_source[g][cell] += C_factor *
